UI 去除派工单模块列表 产品编码 字段

+++ mes/src/main/java/com/panda/pda/mes/discovery/data/WorkOrderDiscoveryModel.kt
@@ -45,5 +45,5 @@ data class WorkOrderDiscoveryModel(
     val updateTime: String,
     val workNo: String,
     val workOrderCode: String,
-    val workOrderNum: Double
+    val workOrderNum: Int
 )
